gracefully handle distributions without source info
Some distributions do not provide source information.
Notably those installed via apt, dnf and probably more
system package managers. Without the source information
we are not able to resolve an import to a distribution.

Instead of failing on such distributions from the start,
we introduce a new warning type. imports which cannot be
resolved to a distribution are labeled with this warning,
which also provides clues on which distributions could not
be resolved.

As an additional benefit, separating these warnings makes it
also easier to surpress them by the user.